FIX/ 코드 정리 + 프로필사진 메뉴 항목 추가
