+++ client/src/components/main/Calendar.js
@@ -1,9 +1,7 @@
-import React, { useEffect, useState, useRef } from "react";
-import { Link } from "react-router-dom";
+import React, { useEffect, useState } from "react";
 import { FullCalendar } from "./FullCalendar";
 import "./_styles.scss";
 import styled from "styled-components";
-import { Icon } from "@iconify/react";
 import DiaryModal from "./WriteModal";
 import Diary from "./Diary";
 import { startOfMonth, endOfMonth, format } from "date-fns";
@@ -12,10 +10,6 @@ import axios from "axios";
 import { useQuery, useMutation, useQueryClient } from "react-query";
 import PacmanLoader from "react-spinners/PacmanLoader";
 import { Triangle } from "react-loader-spinner";
-import { toast, ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-
-// import useMouse from "@react-hook/mouse-position";
 
 const Container = styled.div`
   flex: 1.75;
@@ -52,16 +46,10 @@ export default function Calendar() {
     setShowDiary,
     setShowInstanceTable,
     setcurrentMonth,
-    setPosition,
+    formatter,
   } = calenderStore();
 
   const queryClient = useQueryClient();
-
-  function formatter(e) {
-    let a = e;
-    let string = a[5] + a[6] + "/" + a[8] + a[9] + "/" + a[2] + a[3];
-    return string;
-  }
 
   let stdt = format(startOfMonth(month), "yyyy-MM-dd");
 
@@ -122,9 +110,6 @@ export default function Calendar() {
     },
     {
       onSuccess: () => {
-        // toast.success("일기가 등록 되었습니다!", {
-        //   position: toast.POSITION.TOP_CENTER,
-        // });
         queryClient.invalidateQueries("getCalendarData");
       },
     }
@@ -147,9 +132,6 @@ export default function Calendar() {
     },
     {
       onSuccess: () => {
-        // toast.success("일기가 수정 되었습니다!", {
-        //   position: toast.POSITION.TOP_CENTER,
-        // });
         queryClient.invalidateQueries("getCalendarData");
       },
     }
@@ -172,9 +154,6 @@ export default function Calendar() {
     },
     {
       onSuccess: () => {
-        // toast.success("일기가 삭제 되었습니다!", {
-        //   position: toast.POSITION.TOP_CENTER,
-        // });
         queryClient.invalidateQueries("getCalendarData");
       },
     }
@@ -182,7 +161,6 @@ export default function Calendar() {
 
   useEffect(() => {
     setcurrentMonth(new Date());
-    // setPosition(mouse.y);
     refetch();
   }, []);
 
@@ -208,7 +186,6 @@ export default function Calendar() {
         <PacmanLoader color="#36d7b7" />
       </div>
     );
-  // if (isFetching) return <PacmanLoader color="#36d7b7" />;
   if (isFetching)
     return (
       <div
@@ -219,7 +196,6 @@ export default function Calendar() {
           alignItems: "center",
           width: "100%",
           color: "rgba(109, 73,129, 1)",
-          backgroundColor: "transparent",
         }}
       >
         <Triangle
@@ -231,7 +207,6 @@ export default function Calendar() {
           wrapperClassName=""
           visible={true}
         />
-        {/* <div> 처리중입니다 </div> */}
       </div>
     );
 
@@ -241,10 +216,8 @@ export default function Calendar() {
         setShowInstanceTable(false);
       }}
     >
-      {/* <ToastContainer /> */}
       <FullCalendar
         Y={y}
-        refetch={refetch}
         dateList={data[2]}
         changeValue={changeValue}
         diaryDatas={data[1]}
